refactor(cli): move dependency into its own module
